services/web/requirements.txt
Flask==1.0.2
flask-admin==1.5.2
flask-migrate==2.3.1
flask-security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
gunicorn==19.9.0
psycopg2==2.7.6.1
python-dotenv==0.10.1
